shift model to models, add test for ATmodel
