+++ tests/test_KD_Lib.py
@@ -1,15 +1,15 @@
 # -*- coding: utf-8 -*-
 """Tests for `KD_Lib` package."""
 
-import pytest
-import random
-
+import torch
 from KD_Lib.TAKD.main import main_TAKD
 from KD_Lib.original.mnist import mnist
 from KD_Lib.models.resnet import ResNet18, ResNet34, ResNet50, ResNet101, ResNet152
 
+
 def test_mnist():
     mnist(epochs=0)
+
 
 def test_resnet():
     params = [4,4,8,8,16]
@@ -18,6 +18,7 @@ def test_resnet():
     ResNet50(params)
     ResNet101(params)
     ResNet152(params)
+
 
 def test_TAKD():
     config = {
@@ -58,3 +59,11 @@ def test_TAKD():
         'assistant_train_order': [[-1], [-1, 0]]
     }
     main_TAKD(config)
+
+
+def test_attention_model():
+    params = [4, 4, 8, 8, 16]
+    sample_input = torch.ones(size=(1, 3, 32, 32), requires_grad=False)
+    model = ResNet152(params, att=True)
+    sample_output = model(sample_input)
+    print(sample_output)
